Refactor BrowserWindow code to common util file
Makes it easier to add a new (fetch books metadata) scraper
