+++ src/scraper/index.ts
@@ -1,4 +1,4 @@
-import scrapeHighlightsForBook from './scrapeHighlights';
+import scrapeHighlightsForBook from './scrapeBookHighlights';
 import scrapeBooks from './scrapeBooks';
 import scrapeLogoutUrl from './scrapeLogoutUrl';
